requirements.txt
PyQt5
numpy
rdp
opencv-python-headless
ultralytics
torch
torchvision
onnxruntime
shapely
PyYAML